This command adds the "master" input source, we will use this to show on the frontend.
Currently we have a most credible value for the fields, this way the most credible value (input source ) will always be shown to the user.

This changes now, a new input source will be added (master). This input source will be filled by any other input source.  This will therefore be updated each adjustment on Hoomdossier, the master input source will also be used for the calculations

We will also continue to store the data in the user's own input source (coach / resident).